Tweaked oregen
Certus now also generates in Solara and the Mining Dimension
Thermal Expansion, ProjectRed, and Immersive Engineering ores should all
now generate in Solara and the Mining dimension.
Tried to do the same with Nuclearcraft ores, but it's not working,
despite using the same method, and I have no idea why.